requirements.txt
fastapi
uvicorn
openai
chromadb
langchain
langchain-text-splitters
langchain-core
langchain-community
langchain-groq
langchain-huggingface
sentence-transformers
httpx
beautifulsoup4
requests
pymupdf
aiosqlite
python-dotenv
python-multipart